[Move]SendMessageEffectを作成し、関連メソッドを移動
