+++ src/main/scala/com/github/unchama/seichiassist/subsystems/autosave/bukkit/instances/SyncCanNotifyBukkitSaves.scala
@@ -3,14 +3,14 @@ package com.github.unchama.seichiassist.subsystems.autosave.bukkit.instances
 import cats.effect.Sync
 import com.github.unchama.seichiassist.concurrent.PluginExecutionContexts.onMainThread
 import com.github.unchama.seichiassist.subsystems.autosave.application.CanNotifySaves
-import com.github.unchama.seichiassist.util.Util
+import com.github.unchama.seichiassist.util.SendMessageEffect
 import org.bukkit.Bukkit
 
 object SyncCanNotifyBukkitSaves {
 
   def apply[F[_]: Sync]: CanNotifySaves[F] = (message: String) =>
     Sync[F].delay {
-      Util.sendMessageToEveryoneIgnoringPreference(message)
+      SendMessageEffect.sendMessageToEveryoneIgnoringPreference(message)
       Bukkit.getLogger.info(message)
     }
