+++ src/main/scala/com/github/unchama/seichiassist/subsystems/dragonnighttime/bukkit/instances/SyncNotifiable.scala
@@ -3,13 +3,13 @@ package com.github.unchama.seichiassist.subsystems.dragonnighttime.bukkit.instan
 import cats.effect.Sync
 import com.github.unchama.seichiassist.concurrent.PluginExecutionContexts.onMainThread
 import com.github.unchama.seichiassist.subsystems.dragonnighttime.application.Notifiable
-import com.github.unchama.seichiassist.util.Util
+import com.github.unchama.seichiassist.util.SendMessageEffect
 import org.bukkit.Bukkit
 
 object SyncNotifiable {
   def apply[F[_]: Sync]: Notifiable[F] = (message: String) =>
     Sync[F].delay {
-      Util.sendMessageToEveryoneIgnoringPreference(message)
+      SendMessageEffect.sendMessageToEveryoneIgnoringPreference(message)
       Bukkit.getLogger.info(message)
     }
 }
