+++ src/main/scala/com/github/unchama/seichiassist/task/GiganticBerserkTask.scala
@@ -6,7 +6,7 @@ import com.github.unchama.seichiassist.data.player.PlayerData
 import com.github.unchama.seichiassist.subsystems.discordnotification.DiscordNotificationAPI
 import com.github.unchama.seichiassist.subsystems.mana.ManaApi
 import com.github.unchama.seichiassist.subsystems.mana.domain.ManaAmount
-import com.github.unchama.seichiassist.util.Util
+import com.github.unchama.seichiassist.util.{SendMessageEffect, Util}
 import com.github.unchama.seichiassist.{LevelThresholds, SeichiAssist}
 import org.bukkit.ChatColor._
 import org.bukkit.Sound
@@ -93,7 +93,7 @@ class GiganticBerserkTask {
           DiscordNotificationAPI[F].send(messageWithoutColor).toIO,
           IO {
             Util.sendEverySound(Sound.ENTITY_ENDERDRAGON_DEATH, 1, 1.2f)
-            Util.sendMessageToEveryoneIgnoringPreference(messageWithColor)
+            SendMessageEffect.sendMessageToEveryoneIgnoringPreference(messageWithColor)
           }
         ).sequence
