+++ src/main/scala/com/github/unchama/seichiassist/task/global/PlayerDataBackupRoutine.scala
@@ -4,7 +4,7 @@ import cats.effect.{IO, Timer}
 import com.github.unchama.concurrent.{RepeatingRoutine, RepeatingTaskContext}
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.task.PlayerDataSaveTask
-import com.github.unchama.seichiassist.util.Util
+import com.github.unchama.seichiassist.util.SendMessageEffect
 import org.bukkit.Bukkit
 import org.bukkit.ChatColor._
 
@@ -28,7 +28,7 @@ object PlayerDataBackupRoutine {
 
         for {
           _ <- IO {
-            Util.sendMessageToEveryoneIgnoringPreference(s"${AQUA}プレイヤーデータセーブ中…")
+            SendMessageEffect.sendMessageToEveryoneIgnoringPreference(s"${AQUA}プレイヤーデータセーブ中…")
             Bukkit.getLogger.info(s"${AQUA}プレイヤーデータセーブ中…")
           }
           players <- IO {
@@ -39,7 +39,7 @@ object PlayerDataBackupRoutine {
               .savePlayerData[IO](player, SeichiAssist.playermap(player.getUniqueId))
           }
           _ <- IO {
-            Util.sendMessageToEveryoneIgnoringPreference(s"${AQUA}プレイヤーデータセーブ完了")
+            SendMessageEffect.sendMessageToEveryoneIgnoringPreference(s"${AQUA}プレイヤーデータセーブ完了")
             Bukkit.getLogger.info(s"${AQUA}プレイヤーデータセーブ完了")
           }
         } yield ()
